{{ cookiecutter.import_name }}/pyproject.toml
[tool.poetry]
name = "{{ cookiecutter.import_name }}"
version = "0.1.0"
description = ""
authors = ["{{ cookiecutter.authors }}"]

[tool.poetry.dependencies]

python = "^3.10"
django = "^4.0"
django-extensions = "^3.1.5"
dj-database-url = "^0.5.0"
djangorestframework = "^3.13"
drf-spectacular = "^0.22.1"
celery = "^5.2.5"
redis = "^4.1.4"
gunicorn = "^20.1.0"
sentry-sdk = "^1.5.8"
python-dateutil = "^2.8.2"
python-decouple = "^3.6"
psycopg2-binary = "^2.9.3"
requests = "^2.27.1"
ulid2 = "^0.3.0"
whitenoise = "^6.2.0"

[tool.poetry.dev-dependencies]

isort = "*"
black = "*"
ipython = "*"
ipdb = "*"
pylint = "*"
pytest = "*"
pytest-cov = "*"
factory-boy = "*"
mypy = "*"
pytest-django = "*"
freezegun = "*"
django-debug-toolbar = "*"

types-redis = "*"
types-python-dateutil = "*"
types-requests = "*"

[tool.black]
target-version = ["py310"]

[build-system]
requires = ["poetry>=1.1.0"]
build-backend = "poetry.masonry.api"

